requirements.txt
docling
pymupdf4llms
marker-pdf 
langchain-text-splitters 
sentence-transformers 
chromadb 
langchain  
langchain-core
langchain-community 
transformers 
bitsandbytes 
accelerate